requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
accelerate>=0.33.0
timm>=0.9.8
pillow>=10.3.0
gradio>=4.44.0
safetensors>=0.4.3
numpy>=1.26.4
huggingface-hub>=0.24.5
